updated scripts as per the last versions
