+++ PetcoBDDFramework/src/main/java/stepDefinations/HomePage.java
@@ -1,29 +1,28 @@
 package stepDefinations;
 
 import common.Page_BasePage;
-import cucumber.api.java.After;
-import cucumber.api.java.Before;
-import cucumber.api.java.en.Given;
-import cucumber.api.java.en.Then;
-import cucumber.api.java.en.When;
+import io.cucumber.java.After;
+import io.cucumber.java.Before;
+import io.cucumber.java.en.Given;
+import io.cucumber.java.en.Then;
 import seleniumPages.Page_PetcoHomepage;
 
 public class HomePage {
 	
 	Page_PetcoHomepage petcoHomePage = new Page_PetcoHomepage();
 	Page_BasePage basepage = new Page_BasePage();
-	
-	@Before("@SmokeTest @RegressionTest")
-	 public void setUP()
-	 {
-		basepage.launchPetcoURL();
-	 }
 
-	@After("@SmokeTest @RegressionTest")
-	 public void tearDown()
-	 {
+	@Before
+	public void launchBrowser()
+	{
+		basepage.launchPetcoURL();
+	}
+
+	@After
+	public void closeBrowser()
+	{
 		basepage.closeBrowser();
-	 }
+	}
 	
 	@Given("^user launches Chrome Browser with Petco URL$")
 	public void launch_ChromeBrowser_withPetcoURL()
