--- PetcoBDDFramework/src/main/java/testRunner/testRunner.java
@@ -1,24 +1,0 @@
-package testRunner;
-
-import org.junit.runner.RunWith;
-import cucumber.api.CucumberOptions;
-import cucumber.api.junit.Cucumber;
-
-
-	@RunWith(Cucumber.class)
-	@CucumberOptions(
-			features = "C:/Users/SAI KUMAR/git/Cucumber-BDD-Project/PetcoBDDFramework/src/main/java/Features/AddToCart.feature", //the path of the feature files
-			glue={"stepDefinations"}, //the path of the step definition files
-			format= {"pretty","html:test-outout", "json:json_output/cucumber.json", "junit:junit_xml/cucumber.xml"}, //to generate different types of reporting
-			monochrome = true, //display the console output in a proper readable format
-			strict = true, //it will check if any step is not defined in step definition file
-			dryRun = false, //to check the mapping is proper between feature file and step def file
-			tags = {"@SmokeTest, @RegressionTest"}		
-			)
-
-public class testRunner {
-
-}
-
-	//ORed : tags = {"@SmokeTest , @RegressionTest"} -- execute all tests tagged as @SmokeTest OR @RegressionTest
-	//ANDed : tags = tags = {"@SmokeTest" , "@RegressionTest"} -- execute all tests tagged as @SmokeTest AND @RegressionTest
