+++ PetcoBDDFramework/src/test/java/testRunner/testRunner.java
@@ -1,0 +1,24 @@
+package testRunner;
+
+import org.junit.runner.RunWith;
+import io.cucumber.junit.CucumberOptions;
+import io.cucumber.junit.Cucumber;
+
+
+	@RunWith(Cucumber.class)
+	@CucumberOptions(
+			features = "src/main/resources/Features/login.feature", //the path of the feature files
+			glue={"stepDefinations"}, //the path of the step definition files
+			//format= {"pretty","html:test-outout", "json:json_output/cucumber.json", "junit:junit_xml/cucumber.xml"}, //to generate different types of reporting
+			monochrome = true, //display the console output in a proper readable format
+			//strict = true, //it will check if any step is not defined in step definition file
+			dryRun = false, //to check the mapping is proper between feature file and step def file
+			tags = "@SmokeTest"
+			)
+
+public class testRunner {
+
+}
+
+	//ORed : tags = {"@SmokeTest , @RegressionTest"} -- execute all tests tagged as @SmokeTest OR @RegressionTest
+	//ANDed : tags = tags = {"@SmokeTest" , "@RegressionTest"} -- execute all tests tagged as @SmokeTest AND @RegressionTest
